fix warning
git-svn-id: https://josm.openstreetmap.de/osmsvn/applications/editors/josm/plugins@36341 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ pbf/src/org/openstreetmap/josm/plugins/pbf/io/PbfWriter.java
@@ -149,7 +149,7 @@ public class PbfWriter implements Closeable {
                         b.setUid((int) e.getUser().getId());
                         b.setUserSid(stable.getIndex(e.getUser().getName()));
                     }
-                    b.setTimestamp((int) (e.getInstant().toEpochMilli() / date_granularity));
+                    b.setTimestamp(e.getInstant().toEpochMilli() / date_granularity);
                     b.setVersion(e.getVersion());
                     b.setChangeset(e.getChangesetId());
                 }
